# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "VehiGo API docs",
        "description": "This is an API documentation, generated by L5 Swagger",
        "contact": {
            "email": "office@atelierulmeu.ro"
        },
        "version": "1.0"
    },
    "servers": [
        {
            "url": "/api/v1"
        }
    ],
    "paths": {
...
        "/images": {
            "get": {
                "tags": [
                    "UserImage"
                ],
                "summary": "Get user images",
                "description": "Returns the user images models",
                "operationId": "9ae54cba70b5edcd45d126cca7ecedd5",
                "responses": {
                    "200": {
                        "description": "The user images model response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UserImagesResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer_Auth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "UserImage"
                ],
                "summary": "Upload new user image",
                "description": "Creates a new user image",
                "operationId": "95dcce41710c39ca2ee94dc7a152b5a0",
                "requestBody": {
                    "description": "User image fillable properties",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UserImageBodyUpload"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "User image model response.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UserImageResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer_Auth": []
                    }
                ]
            }
        },
        "/images/{id}": {
            "get": {
                "tags": [
                    "UserImage"
                ],
                "summary": "Get user image by id",
                "description": "Returns the user image. inline/stream",
                "operationId": "5898b0bbe2f8c821c7244e5a4f130905",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id of user image",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": "1"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "The image file response"
                    }
                },
                "security": [
                    {
                        "bearer_Auth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "UserImage"
                ],
                "summary": "Update the user image",
                "description": "Update the user image model. (not the file)",
                "operationId": "9e67f0cc897bc731e906bb75815a896b",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id of user image",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": "1"
                        }
                    }
                ],
                "requestBody": {
                    "description": "User image fillable properties",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UserImageBodyEdit"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "User image model response.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UserImageResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer_Auth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "UserImage"
                ],
                "summary": "Delete user image by id",
                "description": "Delete the user image model and file",
                "operationId": "2e120d717538f7e9026e074bc53f54d0",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id of user image",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": "1"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "The user image deleted successfull",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/DeleteUserImageResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer_Auth": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
...
            "UserImageBodyUpload": {
                "required": [
                    "image"
                ],
                "properties": {
                    "image": {
                        "description": "Multipart form data file",
                        "type": "string",
                        "format": "binary"
                    },
                    "car_id": {
                        "type": "integer",
                        "example": "1"
                    },
                    "favourite": {
                        "type": "boolean",
                        "example": "true"
                    },
                    "visible_name": {
                        "type": "string",
                        "example": "custom name sdfsdf"
                    },
                    "meta_data": {
                        "type": "object",
                        "example": []
                    }
                },
                "type": "object"
            },
            "UserImageBodyEdit": {
                "properties": {
                    "car_id": {
                        "type": "integer",
                        "example": "1"
                    },
                    "favourite": {
                        "type": "boolean",
                        "example": "true"
                    },
                    "visible_name": {
                        "type": "string",
                        "example": "custom name sdfsdf"
                    },
                    "meta_data": {
                        "type": "object",
                        "example": []
                    }
                },
                "type": "object"
            },
            "UserImagesResponse": {
                "properties": {
                    "status": {
                        "type": "string",
                        "example": "success"
                    },
                    "message": {
                        "type": "string",
                        "example": ""
                    },
                    "data": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/UserImage"
                        }
                    }
                },
                "type": "object"
            },
            "UserImageResponse": {
                "properties": {
                    "status": {
                        "type": "string",
                        "example": "success"
                    },
                    "message": {
                        "type": "string",
                        "example": ""
                    },
                    "data": {
                        "$ref": "#/components/schemas/UserImage"
                    }
                },
                "type": "object"
            },
            "DeleteUserImageResponse": {
                "properties": {
                    "status": {
                        "type": "string",
                        "example": "success"
                    },
                    "message": {
                        "type": "string",
                        "example": ""
                    },
                    "data": {
                        "type": "object",
                        "example": []
                    }
                },
                "type": "object"
            },
...
            "UserImage": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "car_id": {
                        "type": "integer",
                        "example": "2"
                    },
                    "name": {
                        "type": "string",
                        "example": "myfirstpicture"
                    },
                    "visible_name": {
                        "type": "string",
                        "example": "custom name sdfsdf"
                    },
                    "type": {
                        "type": "string",
                        "example": "jpg"
                    },
                    "size": {
                        "type": "integer",
                        "example": "2048"
                    },
                    "h_size": {
                        "type": "string",
                        "example": "1 MB"
                    },
                    "favorite": {
                        "type": "boolean",
                        "example": "true"
                    },
                    "meta_data": {
                        "type": "json",
                        "example": []
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "bearer_Auth": {
                "type": "apiKey",
                "description": "A short description for security scheme",
                "name": "Authorization",
                "in": "header"
            }
        }
    }
}